Format menu and presets

+++ src/pages/Index.vue
@@ -1,44 +1,102 @@
 <template>
   <div class="q-pa-md">
-    <div class="q-gutter-md" style="max-width: 80%">
-      <q-input ref="DirectoryInput" v-model="Directory" readonly label="Output Directory" />
-      <q-btn color="secondary" label="Choose directory" @click="ChooseDirectory" />
+    <div class="q-gutter-md">
+      <!--DIRECTORY INPUT-->
+      <q-item clickable v-ripple @click="ChooseDirectory">
+        <q-item-section avatar top>
+          <q-avatar icon="folder" color="primary" text-color="white" />
+        </q-item-section>
+        <q-item-section>
+          <q-item-label>Output Directory</q-item-label>
+          <q-item-label caption>{{Directory}}</q-item-label>
+        </q-item-section>
+      </q-item>
+      <!--URL INPUT-->
       <q-input ref="URLInput" v-model="VideoUrl" label="Video URL" />
+      <!--GET INFO BUTTON-->
       <q-btn
         :loading="IsGettingVideoInformation"
         color="primary"
         label="Get Info"
         @click="GetVideoInfo"
       />
+
     </div>
+
+    <!--VIDEO INFORMATION CARD-->
     <q-card v-if="InfoSection.Show">
-        <q-img :src="InfoSection.data.thumbnail" spinner-color="white" basic class="video-thumbnail">
-          <div class="absolute-bottom text-subtitle2 text-center">{{InfoSection.data.title}}</div>
-        </q-img>
+      <!--THUMBNAIL-->
+      <q-img :src="InfoSection.data.thumbnail" spinner-color="white" basic class="video-thumbnail">
+        <div class="absolute-bottom text-subtitle2 text-center">{{InfoSection.data.title}}</div>
+      </q-img>
       <q-card-section>
-        <div v-for="(Info, index) in InfoSection.InfoToDisplay" :key="`${index}-${Info.field}`">
-          <q-input standout v-model="InfoSection.data[Info.field]" readonly :label="Info.label" />
-        </div>
-        <q-select
-        outlined
-        label="Choose Format"
-        v-model="ChosenFormat"
-        :options="InfoSection.data.formats"
-        :option-value="opt => opt === null ? null : opt"
-        :option-label="opt => opt === null ? '- Null -' : opt.format"
-        map-options
-        />
-        <q-input standout :value="`${ChosenFormat.filesize / 1000000} MB`" readonly label="Size" />
-        <q-input standout :value="ChosenFormat.ext" readonly label="Extension" />
-        <q-input standout :value="ChosenFormat.acodec" readonly label="Audio type" />
-        <q-input standout :value="ChosenFormat.vcodec" readonly label="Video type" />
+        <!--BASIC INFORMATION-->
+        <q-card>
+          <q-card-section v-for="(Info, index) in InfoSection.InfoToDisplay" :key="`${index}-${Info.field}`">
+            <q-item-label>{{Info.label}}</q-item-label>
+            <q-item-label caption lines="2">{{InfoSection.data[Info.field]}}</q-item-label>
+          </q-card-section>
+        </q-card>
+        <q-card>
+          <q-item clickable v-ripple @click="ShowFormatsDialog">
+            <q-item-section>
+              <q-item-label>Format{{ChosenFormat.height ? ` - ${ChosenFormat.height}P` : ''}}</q-item-label>
+              <q-item-label caption>{{ChosenFormat.format}}{{ChosenFormat.filesize ? ' at ' + (Math.round(ChosenFormat.filesize / 1000000)) + 'MB': ''}}</q-item-label>
+            </q-item-section>
+
+            <q-item-section side top>
+              <q-item-label caption>{{ChosenFormat.ext}}</q-item-label>
+            </q-item-section>
+          </q-item>
+        </q-card>
 
       </q-card-section>
+
+      <!--DOWNLOAD BUTTON-->
       <q-card-section>
         <q-btn color="secondary" label="Download" @click="DownloadVideo" />
       </q-card-section>
     </q-card>
 
+    <!--CHOOSE FORMAT DIALOG-->
+    <q-dialog
+    v-model="ShowFormats"
+    transition-show="scale"
+    transition-hide="scale">
+      <q-card style="width: 300px">
+        <q-card-section class="row items-center">
+          <div class="text-h6">Formats</div>
+          <q-space />
+          <q-btn icon="close" flat round dense v-close-popup />
+        </q-card-section>
+        <q-separator />
+        <q-card-section style="max-height: 50vh" class="scroll">
+          <q-list bordered class="rounded-borders">
+            <q-expansion-item
+              v-for="(category, index) in InfoSection.formats" :key="`${index}-Category`"
+              expand-separator
+              :label="`${category.label}`"
+              caption=""
+            >
+            <q-list v-for="(format, index) in category.list" :key="`${index}-format`">
+              <q-item clickable v-close-popup v-ripple @click="ChosenFormat = format" :active="ChosenFormat.format_id === format.format_id">
+                <q-item-section>
+                  <q-item-label>Format{{format.height ? ` - ${format.height}P` : ''}}</q-item-label>
+                  <q-item-label caption>{{format.format}}{{format.filesize ? ' at ' + (Math.round(format.filesize / 1000000)) + 'MB': ''}}</q-item-label>
+                </q-item-section>
+
+                <q-item-section side top>
+                  <q-item-label caption>{{format.ext}}</q-item-label>
+                </q-item-section>
+              </q-item>
+            </q-list>
+            </q-expansion-item>
+          </q-list>
+        </q-card-section>
+      </q-card>
+    </q-dialog>
+
+    <!--DOWNLOAD DIALOG-->
     <q-dialog
       v-model="ShowDownloadDialog"
       persistent
@@ -46,16 +104,21 @@
       transition-hide="scale"
     >
       <q-card style="width: 300px">
+        <!--TITLE AND PROGRESS BAR-->
         <q-card-section>
-          <div class="text-h6">Download Started</div>
-          <q-linear-progress stripe style="height: 15px" :value="this.DownloadSection.progress" />
+          <div class="text-h6">Download in progress</div>
         </q-card-section>
+        <q-separator />
 
+        <!--STATUS-->
         <q-card-section>
           <div v-if="DownloadSection.isFinished" style="color:green">Completed</div>
           <div v-else>{{DownloadSection.status}}</div>
         </q-card-section>
 
+        <q-linear-progress stripe style="height: 15px" :value="this.DownloadSection.progress" />
+
+        <!--DIALOG CLOSE BUTTON-->
         <q-card-actions align="right" class="bg-white text-teal">
           <q-btn flat v-if="DownloadSection.isFinished" label="OK" v-close-popup />
         </q-card-actions>
@@ -79,8 +142,9 @@ export default {
         properties: ['openDirectory']
       })[0];
     },
-    GetVideoInfo: function () {
+    GetVideoInfo: function (isSecondTry) {
       if (!this.VideoUrl) {
+        this.$q.notify('Please input the video url');
         return;
       }
       // Set variables
@@ -88,18 +152,31 @@ export default {
       this.IsGettingVideoInformation = true;
       this.InfoSection.Show = false;
 
+      let args = [];
+      if (!isSecondTry) {
+        args = ['-f', 'bestvideo+bestaudio'];
+      }
       // Begin downloading the video information with the default being bestvideo+bestaudio
-      this.DownloadVideoInfo(this.CurrentVideoUrl, ['-f', 'bestvideo+bestaudio']).then((info) => {
+      this.DownloadVideoInfo(this.CurrentVideoUrl, args).then((info) => {
         // Set variables
         this.IsGettingVideoInformation = false;
         this.InfoSection.data = info;
+        this.InfoSection.formats.audio.list = [];
+        this.InfoSection.formats.video.list = [];
+        this.InfoSection.formats.audioAndVideo.list = [];
+        this.InfoSection.formats.custom.list = [];
 
         // Split format id
         const formatIds = info.format_id.split('+');
         if (formatIds.length > 1) {
           const bestVideo = info.formats.find(format => formatIds.some(formatId => formatId === format.format_id) && format.vcodec !== 'none');
           const bestAudio = info.formats.find(format => formatIds.some(formatId => formatId === format.format_id) && format.acodec !== 'none');
-          const format = {
+
+          const audioOnly = info.formats.filter(format => format.vcodec === 'none');
+          const videoOnly = info.formats.filter(format => format.acodec === 'none');
+          const audioAndVideo = info.formats.filter(format => format.vcodec !== 'none' && format.acodec !== 'none');
+
+          const bestAudioAndVideoPreset = {
             abr: 50,
             acodec: bestAudio.acodec,
             ext: 'mp4',
@@ -111,19 +188,54 @@ export default {
             height: bestVideo.height,
             vcodec: bestVideo.vcodec
           };
-          info.formats.unshift(format);
+          const bestAudioPreset = {
+            abr: 50,
+            acodec: bestAudio.acodec,
+            ext: 'mp3',
+            filesize: bestAudio.filesize,
+            format: 'Best audio',
+            format_id: bestAudio.format_id,
+            format_note: 'Best audio only',
+            quality: -1
+          };
+          const bestVideoPreset = {
+            abr: 50,
+            ext: 'mp4',
+            filesize: bestVideo.filesize,
+            format: 'Best video',
+            format_id: bestVideo.format_id,
+            format_note: 'Best video only',
+            quality: -1,
+            height: bestVideo.height,
+            vcodec: bestVideo.vcodec
+          };
+
+          this.InfoSection.formats.audio.list = audioOnly;
+          this.InfoSection.formats.video.list = videoOnly;
+          this.InfoSection.formats.audioAndVideo.list = audioAndVideo;
+
+          this.InfoSection.formats.custom.list.push(bestAudioAndVideoPreset);
+          this.InfoSection.formats.custom.list.push(bestAudioPreset);
+          this.InfoSection.formats.custom.list.push(bestVideoPreset);
+
+          info.formats.unshift(bestAudioAndVideoPreset);
         }
         this.ChosenFormat = info.formats.find(format => format.format_id === info.format_id);
         this.InfoSection.Show = true;
       }).catch((err) => {
-        console.error(err);
+        if (isSecondTry) {
+          return console.error(err);
+        }
+        this.DownloadVideoInfo(true);
       });
     },
     DownloadVideo: function () {
       if (!this.Directory) {
+        this.$q.notify('Please choose a directory');
         return;
       }
       if (!this.ChosenFormat) {
+        this.$q.notify('Please choose a format');
         return;
       }
 
@@ -133,25 +245,27 @@ export default {
 
       const formatIds = this.ChosenFormat.format_id.split('+');
       const isSplit = formatIds.length > 1;
-      const filename = `${this.InfoSection.data.fulltitle}-${this.ChosenFormat.height}P.${this.ChosenFormat.ext}`.replace('|', '').replace('  ', ' ');
-      let filePath = path.join(this.Directory, filename);
+      const filename = `${this.InfoSection.data.fulltitle}${this.ChosenFormat.height ? ` - ${this.ChosenFormat.height}P` : ''}.${this.ChosenFormat.ext}`.replace('|', '').replace('  ', ' ');
 
+      let filePath = path.join(this.Directory, filename);
       let tempFileNames = [];
       let intervalId;
       let totalDownloaded = [0, 0];
+      let infos = [];
 
-      this.DownloadSection.status = 'Getting video information';
+      this.DownloadSection.status = 'Getting information';
       this.DownloadVideoInfo(this.CurrentVideoUrl, ['-f', formatIds[0]])
         .then((info) => {
           console.log('download info:', info);
+          infos[0] = info;
 
           if (isSplit) {
             tempFileNames[0] = `${info.format_id}.${info.ext}`;
           } else {
-            tempFileNames[0] = `${this.InfoSection.data.fulltitle}-${this.ChosenFormat.height}P.${this.ChosenFormat.ext}`;
+            tempFileNames[0] = `${this.InfoSection.data.fulltitle}${info.height ? ` - ${info.height}P` : ''}.${info.ext}`.replace('|', '').replace('  ', ' '); ;
           }
           filePath = path.join(this.Directory, tempFileNames[0]);
-          this.DownloadSection.status = 'Downloading video';
+          this.DownloadSection.status = 'Downloading';
           intervalId = setInterval(() => {
             fs.stat(filePath, (error, stat) => {
               if (error) return;
@@ -163,6 +277,7 @@ export default {
           return this.StartVideoStream(this.CurrentVideoUrl, this.Directory, tempFileNames[0], info.format_id);
         })
         .then((result) => {
+          if (!isSplit) return;
           console.log(result);
           clearInterval(intervalId);
           if (!isSplit) {
@@ -170,14 +285,17 @@ export default {
           }
           return '';
         }).then(() => {
-          this.DownloadSection.status = 'Getting audio information';
+          if (!isSplit) return;
+          this.DownloadSection.status = 'Getting second part information';
           return this.DownloadVideoInfo(this.CurrentVideoUrl, ['-f', formatIds[1]]);
         }).then((info) => {
+          if (!isSplit) return;
           console.log('Audio info:', info);
+          infos[1] = info;
 
           tempFileNames[1] = `${info.format_id}.${info.ext}`;
           filePath = path.join(this.Directory, tempFileNames[1]);
-          this.DownloadSection.status = 'Downloading audio';
+          this.DownloadSection.status = 'Downloading second part';
           intervalId = setInterval(() => {
             fs.stat(filePath, (error, stat) => {
               if (error) return;
@@ -189,23 +307,43 @@ export default {
 
           return this.StartVideoStream(this.CurrentVideoUrl, this.Directory, tempFileNames[1], info.format_id);
         }).then((result) => {
-          console.log(result);
-          clearInterval(intervalId);
           this.DownloadSection.progress = 0;
-          this.DownloadSection.status = 'Combining audio and video';
+          if (isSplit) {
+            console.log(result);
+            clearInterval(intervalId);
+            this.DownloadSection.status = 'Combining audio and video';
 
-          filePath = path.join(this.Directory, filename);
-          intervalId = setInterval(() => {
-            fs.stat(filePath, (error, stat) => {
-              if (error) return console.error(error);
-              const size = stat.size;
-              this.DownloadSection.progress = size / this.ChosenFormat.filesize;
-            });
-          }, 200);
+            filePath = path.join(this.Directory, filename);
+            intervalId = setInterval(() => {
+              fs.stat(filePath, (error, stat) => {
+                if (error) return console.error(error);
+                const size = stat.size;
+                this.DownloadSection.progress = size / this.ChosenFormat.filesize;
+              });
+            }, 200);
 
-          return this.CombineVideoAndAudio(this.Directory, tempFileNames[0], tempFileNames[1], filename);
-        }).then(() => {
+            return this.CombineVideoAndAudio(this.Directory, tempFileNames[0], tempFileNames[1], filename);
+          } else {
+            if (infos[0].ext !== this.ChosenFormat.ext) {
+              this.DownloadSection.status = 'Converting';
+              filePath = path.join(this.Directory, filename);
+              intervalId = setInterval(() => {
+                fs.stat(filePath, (error, stat) => {
+                  if (error) return console.error(error);
+                  const size = stat.size;
+                  this.DownloadSection.progress = size / this.ChosenFormat.filesize;
+                });
+              }, 200);
+              return this.ConvertFile(this.Directory, tempFileNames[0], filename);
+            }
+          }
+          return true;
+        }).then((skipClean) => {
           clearInterval(intervalId);
+          this.DownloadSection.progress = 1;
+          if (skipClean) {
+            return;
+          }
           tempFileNames.forEach((name) => {
             fs.unlink(path.join(this.Directory, name), (err) => {
               if (err) console.error(err);
@@ -272,24 +410,58 @@ export default {
           resolve(stdout);
         });
       });
+    },
+    ConvertFile: function (directory, input, output) {
+      return new Promise((resolve, reject) => {
+        execFile(path.join(__statics, 'ffmpeg', 'ffmpeg.exe'), ['-i', `"${input}"`, `"${output}"`], {
+          cwd: directory,
+          shell: true
+        },
+        (err, stdout, stderr) => {
+          if (err) return reject(err);
+          resolve(stdout);
+        });
+      });
+    },
+    ShowFormatsDialog: function () {
+      this.ShowFormats = true;
     }
   },
   data () {
     return {
       VideoUrl: '',
       IsGettingVideoInformation: false,
-      Video: {},
       CurrentVideoUrl: '',
       ShowDownloadDialog: false,
       ChosenFormat: {},
       Directory: '',
+      ShowFormats: false,
       InfoSection: {
         Show: false,
         data: {},
+        formats: {
+          custom: {
+            label: 'Preset formats',
+            list: []
+          },
+          audioAndVideo: {
+            label: 'Audio and video formats',
+            list: []
+          },
+          video: {
+            label: 'Video only formats',
+            list: []
+          },
+          audio: {
+            label: 'Audio only formats',
+            list: []
+          }
+        },
         InfoToDisplay: [
           { label: 'Title', field: 'fulltitle' },
           { label: 'Uploader', field: 'uploader' },
-          { label: 'Duration', field: 'duration_formatted' }
+          { label: 'Duration', field: 'duration_formatted' },
+          { label: 'Website', field: 'extractor_key' }
         ]
       },
       DownloadSection: {
@@ -308,5 +480,8 @@ export default {
     max-width: 500px;
     display: block;
     margin: 0 auto;
+  }
+  .directory{
+    display: inline;
   }
 </style>
